store_backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.35.59
botocore==1.35.59
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.3
cryptography==44.0.2
Django==5.1.1
django-cors-headers==4.5.0
django-environ==0.12.0
django-modeltranslation==0.19.10
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
ecdsa==0.19.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
django-cloudinary-storage==0.3.0
stripe==12.0.0
dj-database-url==2.3.0
